feat(payment): create payment gateway midtrans

+++ apps/api/src/models/orders.model.js
@@ -23,7 +23,7 @@ export const init = (sequelize) => {
       userAddressId: DataTypes.INTEGER,
       warehouseId: DataTypes.INTEGER,
       totalPrice: DataTypes.DECIMAL,
-      totalQuantity: DataTypes.DECIMAL,
+      totalQuantity: DataTypes.INTEGER,
       shippingCost: DataTypes.DECIMAL,
       orderDate: DataTypes.DATE,
       orderStatusId: DataTypes.INTEGER,
